refactor (schedulers/time): breakdown large file of scheduler into smaller ones
